refactor(trusty-common): split chat_sessions.rs into submodules under 500-SLOC cap (#607)
Replaces the 510-SLOC chat_sessions.rs monolith with a chat_sessions/
directory module:

  chat_sessions/
    mod.rs   — thin re-export facade + all tests (~120 SLOC)
    types.rs — ChatMessage, ChatSession, ChatSessionMeta, ChatSessionRecord,
               ChatSessionStoreError, Result alias, parse_timestamp (~120 SLOC)
    store.rs — ChatSessionStore open/CRUD operations (~220 SLOC)

Zero API change: all previously pub items remain pub; the public surface
exposed by memory_core/store/mod.rs is unchanged.
